fix: check is user login
