+++ src/pages/FollowersPage.jsx
@@ -14,13 +14,16 @@ import NavBar from "components/NavBar";
 import SideBar from "components/SideBar";
 import SwitchBar from "components/SwitchBar";
 import UserItem from "components/UserItem";
+import { InfoContext } from "contexts/InfoContext";
 
 function FollowersPage() {
   const [currentPage, setCurrentPage] = useState("followers");
   const [followers, setFollowers] = useState([]);
-  const { isTweetModalShow, handleTweetClick } = useContext(TweetContext);
-  const navigate = useNavigate();
 
+  const { isTweetModalShow, handleTweetClick } = useContext(TweetContext);
+  const { isUserLogin, loginAlert } = useContext(InfoContext);
+
+  const navigate = useNavigate();
   const location = useLocation();
   const pageUserId = Number(location.pathname.split("/")[2]);
 
@@ -59,6 +62,16 @@ function FollowersPage() {
     }
   };
 
+  // useEffect
+  // 驗證登入
+  useEffect(() => {
+    if (!isUserLogin) {
+      loginAlert();
+      navigate("/login");
+    }
+  }, [isUserLogin, loginAlert, navigate]);
+
+  // 取得當前使用者的追隨者列表
   useEffect(() => {
     const getFollowersAsync = async () => {
       try {
